+++ language/src/main/java/i4gl/nodes/variables/read/ReadRecordFieldNode.java
@@ -6,6 +6,7 @@ import com.oracle.truffle.api.dsl.Specialization;
 import com.oracle.truffle.api.instrumentation.StandardTags.ReadVariableTag;
 import com.oracle.truffle.api.instrumentation.Tag;
 
+import i4gl.exceptions.UnexpectedRuntimeException;
 import i4gl.nodes.expression.ExpressionNode;
 import i4gl.runtime.types.BaseType;
 import i4gl.runtime.values.Record;
@@ -13,9 +14,10 @@ import i4gl.runtime.values.Record;
 @NodeChild(value = "record", type = ExpressionNode.class)
 @NodeField(name = "identifier", type = String.class)
 @NodeField(name = "fieldType", type = BaseType.class)
-public abstract class ReadFromRecordNode extends ExpressionNode {
+public abstract class ReadRecordFieldNode extends ExpressionNode {
 
     protected abstract BaseType getFieldType();
+
     protected abstract String getIdentifier();
 
     @Override
@@ -23,38 +25,38 @@ public abstract class ReadFromRecordNode extends ExpressionNode {
         return this.getFieldType();
     }
 
-    @Specialization(guards = "returnsChar()")
-    protected char readChar(final Record record) {
-        return (char)record.getCharSafe(getIdentifier());
+    @Specialization(guards = "record.isChar(getIdentifier())")
+    protected char readChar1(final Record record) {
+        throw new UnexpectedRuntimeException();
+        // return record.getCharSafe(getIdentifier());
     }
 
-    @Specialization(guards = "returnsSmallInt()")
+    @Specialization(guards = "record.isSmallInt(getIdentifier())")
     protected short readSmallInt(final Record record) {
-        Short value = record.getSmallIntSafe(getIdentifier());
-        return value.shortValue();
+        return record.getSmallIntSafe(getIdentifier());
     }
 
-    @Specialization(guards = "returnsInt()")
+    @Specialization(guards = "record.isInt(getIdentifier())")
     protected int readInt(final Record record) {
         return record.getIntSafe(getIdentifier());
     }
 
-    @Specialization(guards = "returnsBigInt()")
+    @Specialization(guards = "record.isBigInt(getIdentifier())")
     protected long readBigInt(final Record record) {
         return record.getBigIntSafe(getIdentifier());
     }
 
-    @Specialization(guards = "returnsSmallFloat()")
+    @Specialization(guards = "record.isSmallFloat(getIdentifier())")
     protected float readSmallFloat(final Record record) {
         return record.getSmallFloatSafe(getIdentifier());
     }
 
-    @Specialization(guards = "returnsFloat()")
+    @Specialization(guards = "record.isFloat(getIdentifier())")
     protected double readFloat(final Record record) {
         return record.getFloatSafe(getIdentifier());
     }
 
-    @Specialization(replaces = { "readChar", "readInt", "readBigInt", "readSmallFloat", "readFloat" })
+    @Specialization(replaces = { "readChar1", "readSmallInt", "readInt", "readBigInt", "readSmallFloat", "readFloat" })
     protected Object readObject(final Record record) {
         return record.getObject(getIdentifier());
     }