+++ language/src/main/java/i4gl/runtime/types/complex/CursorType.java
@@ -4,8 +4,8 @@ import com.oracle.truffle.api.CompilerAsserts;
 import com.oracle.truffle.api.frame.FrameSlotKind;
 import com.oracle.truffle.api.interop.InteropLibrary;
 
-import i4gl.exceptions.I4GLRuntimeException;
 import i4gl.exceptions.NotImplementedException;
+import i4gl.exceptions.ShouldNotReachHereException;
 import i4gl.runtime.types.BaseType;
 import i4gl.runtime.values.Cursor;
 
@@ -47,6 +47,6 @@ public class CursorType extends BaseType {
 
     @Override
     public String getNullString() {
-        throw new I4GLRuntimeException("Should not be here");
+        throw new ShouldNotReachHereException();
     }
 }