+++ language/src/main/java/i4gl/runtime/types/compound/RecordType.java
@@ -7,7 +7,7 @@ import com.oracle.truffle.api.CompilerAsserts;
 import com.oracle.truffle.api.frame.FrameSlotKind;
 import com.oracle.truffle.api.interop.InteropLibrary;
 
-import i4gl.exceptions.I4GLRuntimeException;
+import i4gl.exceptions.ShouldNotReachHereException;
 import i4gl.runtime.types.BaseType;
 import i4gl.runtime.types.primitive.IntType;
 import i4gl.runtime.values.Record;
@@ -99,6 +99,6 @@ public class RecordType extends BaseType {
 
     @Override
     public String getNullString() {
-        throw new I4GLRuntimeException("Should not be here");
+        throw new ShouldNotReachHereException();
     }
 }